+++ services/observability/v1api/model_plan_model.go
@@ -12,7 +12,6 @@ Contact: stackit-argus@mail.schwarz
 package v1api
 
 import (
-	"bytes"
 	"encoding/json"
 	"fmt"
 )
@@ -41,6 +40,7 @@ type PlanModel struct {
 	TargetNumber            int32    `json:"targetNumber"`
 	TotalMetricSamples      int32    `json:"totalMetricSamples"`
 	TracesStorage           int32    `json:"tracesStorage"`
+	AdditionalProperties    map[string]interface{}
 }
 
 type _PlanModel PlanModel
@@ -593,6 +593,11 @@ func (o PlanModel) ToMap() (map[string]interface{}, error) {
 	toSerialize["targetNumber"] = o.TargetNumber
 	toSerialize["totalMetricSamples"] = o.TotalMetricSamples
 	toSerialize["tracesStorage"] = o.TracesStorage
+
+	for key, value := range o.AdditionalProperties {
+		toSerialize[key] = value
+	}
+
 	return toSerialize, nil
 }
 
@@ -635,15 +640,38 @@ func (o *PlanModel) UnmarshalJSON(data []byte) (err error) {
 
 	varPlanModel := _PlanModel{}
 
-	decoder := json.NewDecoder(bytes.NewReader(data))
-	decoder.DisallowUnknownFields()
-	err = decoder.Decode(&varPlanModel)
+	err = json.Unmarshal(data, &varPlanModel)
 
 	if err != nil {
 		return err
 	}
 
 	*o = PlanModel(varPlanModel)
+
+	additionalProperties := make(map[string]interface{})
+
+	if err = json.Unmarshal(data, &additionalProperties); err == nil {
+		delete(additionalProperties, "alertMatchers")
+		delete(additionalProperties, "alertReceivers")
+		delete(additionalProperties, "alertRules")
+		delete(additionalProperties, "amount")
+		delete(additionalProperties, "bucketSize")
+		delete(additionalProperties, "description")
+		delete(additionalProperties, "grafanaGlobalDashboards")
+		delete(additionalProperties, "grafanaGlobalOrgs")
+		delete(additionalProperties, "grafanaGlobalSessions")
+		delete(additionalProperties, "grafanaGlobalUsers")
+		delete(additionalProperties, "id")
+		delete(additionalProperties, "logsAlert")
+		delete(additionalProperties, "logsStorage")
+		delete(additionalProperties, "name")
+		delete(additionalProperties, "planId")
+		delete(additionalProperties, "samplesPerScrape")
+		delete(additionalProperties, "targetNumber")
+		delete(additionalProperties, "totalMetricSamples")
+		delete(additionalProperties, "tracesStorage")
+		o.AdditionalProperties = additionalProperties
+	}
 
 	return err
 }