+++ services/observability/v1api/model_update_scrape_config_payload_static_configs_inner.go
@@ -12,7 +12,6 @@ Contact: stackit-argus@mail.schwarz
 package v1api
 
 import (
-	"bytes"
 	"encoding/json"
 	"fmt"
 )
@@ -25,7 +24,8 @@ type UpdateScrapeConfigPayloadStaticConfigsInner struct {
 	// Labels assigned to all metrics scraped from the targets. `Additional Validators:` * should not contain more than 5 keys * each key and value should not be longer than 200 characters
 	Labels map[string]interface{} `json:"labels,omitempty"`
 	// The targets specified by the static config.
-	Targets []string `json:"targets"`
+	Targets              []string `json:"targets"`
+	AdditionalProperties map[string]interface{}
 }
 
 type _UpdateScrapeConfigPayloadStaticConfigsInner UpdateScrapeConfigPayloadStaticConfigsInner
@@ -118,6 +118,11 @@ func (o UpdateScrapeConfigPayloadStaticConfigsInner) ToMap() (map[string]interfa
 		toSerialize["labels"] = o.Labels
 	}
 	toSerialize["targets"] = o.Targets
+
+	for key, value := range o.AdditionalProperties {
+		toSerialize[key] = value
+	}
+
 	return toSerialize, nil
 }
 
@@ -145,15 +150,21 @@ func (o *UpdateScrapeConfigPayloadStaticConfigsInner) UnmarshalJSON(data []byte)
 
 	varUpdateScrapeConfigPayloadStaticConfigsInner := _UpdateScrapeConfigPayloadStaticConfigsInner{}
 
-	decoder := json.NewDecoder(bytes.NewReader(data))
-	decoder.DisallowUnknownFields()
-	err = decoder.Decode(&varUpdateScrapeConfigPayloadStaticConfigsInner)
+	err = json.Unmarshal(data, &varUpdateScrapeConfigPayloadStaticConfigsInner)
 
 	if err != nil {
 		return err
 	}
 
 	*o = UpdateScrapeConfigPayloadStaticConfigsInner(varUpdateScrapeConfigPayloadStaticConfigsInner)
+
+	additionalProperties := make(map[string]interface{})
+
+	if err = json.Unmarshal(data, &additionalProperties); err == nil {
+		delete(additionalProperties, "labels")
+		delete(additionalProperties, "targets")
+		o.AdditionalProperties = additionalProperties
+	}
 
 	return err
 }