+++ services/scf/v1api/model_org_manager_delete_response.go
@@ -12,7 +12,6 @@ Contact: support@stackit.cloud
 package v1api
 
 import (
-	"bytes"
 	"encoding/json"
 	"fmt"
 )
@@ -22,7 +21,8 @@ var _ MappedNullable = &OrgManagerDeleteResponse{}
 
 // OrgManagerDeleteResponse struct for OrgManagerDeleteResponse
 type OrgManagerDeleteResponse struct {
-	Message string `json:"message" validate:"regexp=^[a-zA-Z0-9\\/_]+$"`
+	Message              string `json:"message" validate:"regexp=^[a-zA-Z0-9\\/_]+$"`
+	AdditionalProperties map[string]interface{}
 }
 
 type _OrgManagerDeleteResponse OrgManagerDeleteResponse
@@ -80,6 +80,11 @@ func (o OrgManagerDeleteResponse) MarshalJSON() ([]byte, error) {
 func (o OrgManagerDeleteResponse) ToMap() (map[string]interface{}, error) {
 	toSerialize := map[string]interface{}{}
 	toSerialize["message"] = o.Message
+
+	for key, value := range o.AdditionalProperties {
+		toSerialize[key] = value
+	}
+
 	return toSerialize, nil
 }
 
@@ -107,15 +112,20 @@ func (o *OrgManagerDeleteResponse) UnmarshalJSON(data []byte) (err error) {
 
 	varOrgManagerDeleteResponse := _OrgManagerDeleteResponse{}
 
-	decoder := json.NewDecoder(bytes.NewReader(data))
-	decoder.DisallowUnknownFields()
-	err = decoder.Decode(&varOrgManagerDeleteResponse)
+	err = json.Unmarshal(data, &varOrgManagerDeleteResponse)
 
 	if err != nil {
 		return err
 	}
 
 	*o = OrgManagerDeleteResponse(varOrgManagerDeleteResponse)
+
+	additionalProperties := make(map[string]interface{})
+
+	if err = json.Unmarshal(data, &additionalProperties); err == nil {
+		delete(additionalProperties, "message")
+		o.AdditionalProperties = additionalProperties
+	}
 
 	return err
 }